Ugh why am I so bad at coding

+++ src/python/runners/bandit/successive_halving.py
@@ -43,7 +43,7 @@ class FiniteSuccessiveHalvingRunner:
                       for i in xrange(self.num_arms)
                       if self.still_pull[i]}
             num_arms_i = self.num_arms * self.eta**(-i)
-            r_i = self.inner_num_rounds * self.eta**(i)
+            r_i = int(self.inner_num_rounds * self.eta**(i))
             num_processes_i = int(min([12, num_arms_i]))
 
             print '\tRunning inner loop for', r_i, 'rounds'
